pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simulateinput"
version = "0.1.0"
description = "Cross-platform desktop and browser automation test platform."
readme = "docs/automation-platform-design.md"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "pytesseract>=0.3.13",
    "pywinauto>=0.6.9",
    "pyobjc-framework-Quartz>=11.1; platform_system=='Darwin'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
simctl = "simulateinput.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
